Camera settings

+++ Assets/Scripts/UI/DisplayValue.cs
@@ -4,5 +4,5 @@ using UnityEngine;
 [RequireComponent(typeof(TextMeshProUGUI))]
 public class DisplayValue : MonoBehaviour
 {
-    public void SetText(int amount) => GetComponent<TextMeshProUGUI>().text = amount.ToString();
+    public void SetText(float amount) => GetComponent<TextMeshProUGUI>().text = amount.ToString();
 }
